requirements.txt
openai
pandas
numpy
python-dotenv
ipykernel
langchain
chromadb
tiktoken
streamlit
pypdf2
faiss-cpu
huggingface_hub